remade fsm to be more like fsm, moved conditional logic into states
